requirements.txt
pefile>=2021.5.24
flare-capa>=2.0.0
yara-python>=4.1.2
valhallaAPI>=0.5.2
